+++ crypto-app/jsconfig.json
@@ -1,0 +1,7 @@
+
+{
+  "compilerOptions": {
+    "baseUrl": "src"
+  },
+  "include": ["src"]
+}
