--- crypto-app/src/App.css
@@ -1,74 +1,0 @@
-.table-container{
-  margin: 22px;
-}
-
-table{
-  width: 100%;
-  counter-reset: rowNumber;
-  border-collapse: collapse ;
-  
-}
-
-
-
-table, th, tr, td {
-  border-bottom: 1px solid black;
-  border-left: none;
-  padding: 15px;
-}
-
-
-
-table .data-row::before{
-    display: table-cell;
-    counter-increment: rowNumber;
-    content: counter(rowNumber);
-    min-width: 1em;
-    padding: 0px 80px, 0px, 10px;
-    vertical-align: middle;
-}
-
-.row-number{
-  text-align: left;
-  
-}
-
-th {
-  text-align:start;
-}
-
-td {
-  padding: 20px 0px 20px 0px;
-  
-  
-}
-
-.coin-image {
-  width: 40px;
-  vertical-align: middle;
-  margin-right: 5px;
-}
-
-.percent-red {
-  color: red;
-}
-
-.percent-green {
-  color: limegreen;
-}
-
-.caret-icon {
-  font-size: 12px;
-}
-
-.table-chart{
-  width: 116.93px;
-  height: 40px;
-  
- 
-  
-}
-
-.chart-padding{
-  padding-right: 17px;
-}
