+++ crypto-app/src/components/CoinInstance/index.js
@@ -1,0 +1,1 @@
+export {default} from './CoinInstance'
