+++ crypto-app/src/components/ProgressBar/ProgressBar.js
@@ -1,0 +1,7 @@
+import {Progress} from './styles'
+
+const ProgressBar = (props) => {
+  return <Progress value={props.progress} max="100" />;
+};
+
+export default ProgressBar;
