+++ crypto-app/src/components/ProgressBar/index.js
@@ -1,0 +1,1 @@
+export {default} from './ProgressBar'
