+++ crypto-app/src/components/ProgressBar/styles.js
@@ -1,0 +1,6 @@
+import styled from 'styled-components';
+
+export const Progress = styled.progress`
+  width: 190px;
+  height: 10;
+`;
