+++ crypto-app/src/components/TableCharts/index.js
@@ -1,0 +1,1 @@
+export {default} from './TableCharts'
