+++ crypto-app/src/components/TableCharts/styles.js
@@ -1,0 +1,6 @@
+import styled from 'styled-components'
+
+export const ChartTable = styled.div`
+  width: 116.93px;
+  height: 40px;
+`;
