+++ crypto-app/src/components/index.js
@@ -1,0 +1,3 @@
+export {default as CoinInstance} from './CoinInstance'
+export { default as TableCharts } from './TableCharts';
+export { default as ProgressBar} from './ProgressBar';
