--- crypto-app/src/componets/TableCharts/labelsarr.js
@@ -1,58 +1,0 @@
-export const labels = [
-  "3",
-  "6",
-  "9",
-  "12",
-  "15",
-  "18",
-  "21",
-  "24",
-  "3",
-  "6",
-  "9",
-  "12",
-  "15",
-  "18",
-  "21",
-  "24",
-  "3",
-  "6",
-  "9",
-  "12",
-  "15",
-  "18",
-  "21",
-  "24",
-  "3",
-  "6",
-  "9",
-  "12",
-  "15",
-  "18",
-  "21",
-  "24",
-  "3",
-  "6",
-  "9",
-  "12",
-  "15",
-  "18",
-  "21",
-  "24",
-  "3",
-  "6",
-  "9",
-  "12",
-  "15",
-  "18",
-  "21",
-  "24",
-  "3",
-  "6",
-  "9",
-  "12",
-  "15",
-  "18",
-  "21",
-  "24"
-];
