+++ crypto-app/src/pages/CoinPage/index.js
@@ -1,0 +1,1 @@
+export {default} from './CoinPage'
