+++ crypto-app/src/pages/Portfolio/Portfolio.js
@@ -1,0 +1,7 @@
+import React from "react";
+
+const Portfolio = (props) => (
+    <>hi</>
+)
+
+export default Portfolio;
