+++ crypto-app/src/pages/Portfolio/index.js
@@ -1,7 +1,1 @@
-import React from "react";
-
-const Portfolio = (props) => (
-    <>hi</>
-)
-
-export default Portfolio;
+export {default} from './Portfolio'
