+++ crypto-app/src/pages/index.js
@@ -1,0 +1,3 @@
+export {default as CoinPage} from './CoinPage';
+export {default as Coins} from './Coins';
+export {default as Portfolio} from './Portfolio';
